Add settings generator webpage
Created a basic .html file that can be used to generate xml to put into
the settings.xml file when configuring the Manage Creature Spawns mod to work. This is just a basic, first-round implementation.  I expect there will still be some bugs with it, but I'm crowd sourcing the testing for my sake.